pyproject.toml
[project]
name = "pyfastmart"
version = "0.1.0"
description = "Minimal cart implementation with Stripe integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "stripe>=8.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "toml>=0.10.2",
]

[project.scripts]
pfm = "pfm.__main__:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pfm/"]
